fix: Award 내 prize Long 타입으로 변경
정말 적은 확률로 1등이 2번 당첨될 수 있으니, 상금 금액은 Long 타입으로 변경하는 것은 어떨까요?

_Originally posted by @malibinYun in https://github.com/next-step/kotlin-lotto/pull/480#discussion_r1031884403_

+++ src/main/kotlin/lotto/domain/Award.kt
@@ -1,7 +1,7 @@
 package lotto.domain
 
 enum class Award(
-    val prize: Int,
+    val prize: Long,
     val matchCount: Int,
 ) {
     NON_PLACE(0, 0),
